Add files via upload
I thought it would be cool to create a personal assistant in Python. If you are into movies you may have heard of Jarvis, an A.I. based character in the Iron Man films. 

In this tutorial we will create a robot.


What can this A.I. assistant do for you?

It can send emails on your behalf.

It can play music for you.

It can do Wikipedia searches for you.

It is capable of opening websites like Google, Youtube, etc., in a web browser.

It is capable of opening your code editor or IDE with a single voice command.

Enough talks! Let's start building our own J.A.R.V.I.S